+++ ScrollableLobbyUI/InLobbyConfigIntegration.cs
@@ -33,6 +33,7 @@ namespace ScrollableLobbyUI
                         new BooleanConfigField(ScrollableLobbyUIPlugin.InlinePageArrows.Definition.Key, ScrollableLobbyUIPlugin.InlinePageArrows.Description.Description, () => ScrollableLobbyUIPlugin.InlinePageArrows.Value, (newValue) => ScrollableLobbyUIPlugin.InlinePageArrows.Value = newValue),
                         new BooleanConfigField(ScrollableLobbyUIPlugin.ScrollRowsWithMouseWheel.Definition.Key, ScrollableLobbyUIPlugin.ScrollRowsWithMouseWheel.Description.Description, () => ScrollableLobbyUIPlugin.ScrollRowsWithMouseWheel.Value, (newValue) => ScrollableLobbyUIPlugin.ScrollRowsWithMouseWheel.Value = newValue),
                         new IntConfigField(ScrollableLobbyUIPlugin.SkillRowHeight.Definition.Key, ScrollableLobbyUIPlugin.SkillRowHeight.Description.Description, () => ScrollableLobbyUIPlugin.SkillRowHeight.Value, null, (newValue) => ScrollableLobbyUIPlugin.SkillRowHeight.Value = newValue, 1, 320),
+                        new IntConfigField(ScrollableLobbyUIPlugin.SkillCountForGridButton.Definition.Key, ScrollableLobbyUIPlugin.SkillCountForGridButton.Description.Description, () => ScrollableLobbyUIPlugin.SkillCountForGridButton.Value, null, (newValue) => ScrollableLobbyUIPlugin.SkillCountForGridButton.Value = newValue, 1, int.MaxValue),
                     }
                 }
             };