+++ ScrollableLobbyUI/SkillGridButton.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using RoR2.UI;
+using UnityEngine;
+
+namespace ScrollableLobbyUI
+{
+    public class SkillGridButton : MonoBehaviour
+    {
+        public HGButton button;
+    }
+}